Overlay: fix OS X pid comparison.
Cast our queried pid to quint64 before
comparing.

+++ src/mumble/Overlay.cpp
@@ -187,7 +187,7 @@ void Overlay::toggleShow() {
 				ProcessSerialNumber psn;
 				GetFrontProcess(&psn);
 				GetProcessPID(&psn, &pid);
-				if (pid != oc->uiPid)
+				if (static_cast<quint64>(pid) != oc->uiPid)
 					continue;
 #if 0
 				// Fullscreen only.
